Fix sonar bugs (#236)
* Fix sonar bugs

* Add more tests

* Add more tests

+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/serialization/binary/ReadableByteChannelReader.java
@@ -141,7 +141,7 @@ class ReadableByteChannelReader extends AbstractBinaryReader
 
     private void fillBuffer(ByteBuffer buffer)
     {
-        int start = buffer.position();
+        long start = buffer.position();
         int read;
         while (buffer.hasRemaining())
         {
